// src/app/models/eval.model.ts
export class Eval{
   
        Nom:string=""
        Prenom: string="";
        poste: string="";
        CIN:string="";
        Email:string="";
        objClair: string="";
        objAtteint:string="";
        contenuAd: string="";
        appContenu: string="";
        dureeSession:string="";
        horairePrevu:string="";
        motiviation:string="";
        tHemeFor: string="";
        homogeneite: string="";
        niveau: string="";
        animationGroupe: string="";
        supportPe: string="";
        presentationInf: string="";
        maitriseLing : string="";
        formateurQues : string="";
       formateurPresent : string="";
        formateurGestuel : string="";
        formateurPsent : string="";
        formateurSerieux : string="";
       entenduF: string="";
       etoile: string="";
       role :string="";

}